Organize source into package
