+++ test/test_bandit.py
@@ -1,7 +1,4 @@
-import os
-import sys
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'src')))
-import bandit
+from webninja import bandit
 
 
 def test_tor_driver():
