+++ test/test_surfer.py
@@ -1,7 +1,4 @@
-import os
-import sys
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'src')))
-from surfer import get_domain
+from webninja.surfer import get_domain
 
 
 def test_get_domain():
