+++ webninja/constant.py
