+++ webninja/new.py
