+++ webninja/reader.py
